pyproject.toml
[project]
name = "hmcpl-library"
version = "0.1.0"
description = "CLI for Huntsville-Madison County Public Library account management"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]

[project.scripts]
hmcpl = "hmcpl.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hmcpl"]
